% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_example.R
\name{create_example_qprojektreport}
\alias{create_example_qprojektreport}
\title{Create Example for German Version of Project Report of Qualitas AG}
\usage{
create_example_qprojektreport(ps_path, pb_open = rlang::is_interactive())
}
\arguments{
\item{ps_path}{target path where example report should be positioned}

\item{pb_open}{should the example report source file be opened}
}
\description{
A prepared example report consisting of an RMarkdown source file, a logo and
a diagram is copied to a target directory given by the parameter ps_path.
}
\examples{
tmprepdir <- tempdir()
create_example_qprojektreport(ps_path = tmprepdir, pb_open = FALSE)
unlink(file.path(tmprepdir, "qprojekt_report"), recursive = TRUE, force = TRUE)
}
