% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_rmd.R
\name{draft_qbdpdf2}
\alias{draft_qbdpdf2}
\title{Create Bookdown PDF Document}
\usage{
draft_qbdpdf2(
  ps_path,
  ps_template = "qbdpdf2",
  ps_package = "qrmdtmpl",
  ps_create_dir = "default",
  pb_open = rlang::is_interactive(),
  pl_repl_value = NULL
)
}
\arguments{
\item{ps_path}{path to newly created document}

\item{ps_template}{template to be used}

\item{ps_package}{package where template is stored}

\item{ps_create_dir}{option whether to create new directory for new document}

\item{pb_open}{whether document should be opened}

\item{pl_repl_value}{list with pattern replacement values}
}
\description{
Wrapper to create bookdown pdf document using \code{rmarkdown::draft}.
}
\examples{
\donotrun{
draft_qbdpdf2(ps_path = "example_bdpdf2")
}
}
