% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_rmd.R
\name{draft_qgenericdoc}
\alias{draft_qgenericdoc}
\title{Create Generic Qualitas Document}
\usage{
draft_qgenericdoc(
  ps_path,
  ps_template = "qgenericdoc",
  ps_package = "qrmdtmpl",
  ps_create_dir = "default",
  pb_open = rlang::is_interactive(),
  pl_repl_values = NULL
)
}
\arguments{
\item{ps_path}{path to the document to be created}

\item{ps_template}{name of the tempalte}

\item{ps_package}{package which contains the template}

\item{ps_create_dir}{specify whether to create a new directory}

\item{pb_open}{open the created file in rstudio editor}

\item{pl_repl_values}{list with values to replace placeholders}
}
\value{
invisible(TRUE)
}
\description{
Wrapper function for rmarkdown::draft to create an empty Qualitas AG
Rmarkdown document including the logo on the title page.
}
\details{
The template 'qemptydoc' is taken from this package 'qrmdtmpl'.
}
\examples{
\dontrun{
draft_qgenericdoc(ps_path = 'empty_test_doc')
}
}
