% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_rmd.R
\name{draft_qprojektreport}
\alias{draft_qprojektreport}
\title{Create Qualitas FB ZWS Report (German)}
\usage{
draft_qprojektreport(
  ps_path,
  ps_template = "qprojektreport",
  ps_package = "qrmdtmpl",
  ps_create_dir = "default",
  pb_edit = FALSE,
  pb_open = rlang::is_interactive()
)
}
\arguments{
\item{ps_path}{path to the document to be created}

\item{ps_template}{name of the tempalte}

\item{ps_package}{package which contains the template}

\item{ps_create_dir}{specify whether to create a new directory}

\item{pb_edit}{flag whether newly created file should be edited}

\item{pb_open}{open the created file in rstudio editor}
}
\value{
invisible(TRUE)
}
\description{
Wrapper function for rmarkdown::draft to create a skeleton for a Qualitas AG
report document in German. The document includes the logo on the title page.
}
\details{
The template 'qprojektreport' is taken from this package 'qrmdtmpl'.
}
\examples{
\dontrun{
draft_qprojektreport(ps_path = 'empty_test_doc')
}
}
