% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_rmd.R
\name{draft_qtufte}
\alias{draft_qtufte}
\title{Create Draft Version of a Tufte Report}
\usage{
draft_qtufte(
  ps_path,
  ps_template = "qtufte",
  ps_package = "qrmdtmpl",
  ps_create_dir = "default",
  pb_open = rlang::is_interactive(),
  pl_repl_values = NULL
)
}
\arguments{
\item{ps_path}{path to the document to be created}

\item{ps_template}{name of the tempalte}

\item{ps_package}{package which contains the template}

\item{ps_create_dir}{specify whether to create a new directory}

\item{pb_open}{open the created file in rstudio editor}

\item{pl_repl_values}{list with values to replace placeholders}
}
\description{
Wrapper function to create a draft of a Tufte Report using
\code{rmarkdown::draft()}.
}
\details{
The used RMarkdown template qtufte is taken from this package qrmdtmpl.
}
\examples{
\dontrun{
draft_qtufte(ps_path = "qtufte_report")
}
}
