% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_helper.R
\name{extend_repl_values_qtufte}
\alias{extend_repl_values_qtufte}
\title{Extend Replacement Value List}
\usage{
extend_repl_values_qtufte(pl_repl_values)
}
\arguments{
\item{pl_repl_values}{given value replacement list}
}
\value{
l_repl_value_result extended value replacement list
}
\description{
A given value replacement list is extended by a reasonable choice of defaults
for all missing list elements. The defaults are obtained by special getter
function.
}
\examples{
\dontrun{
extend_repl_values_qtufte(pl_repl_values = list(title = "Tufte Report Title"))
}
}
