% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_rmd.R
\name{generic_rmd_draft}
\alias{generic_rmd_draft}
\title{Generic Document Draft Wrapper Function}
\usage{
generic_rmd_draft(
  ps_path,
  ps_template,
  ps_package,
  ps_create_dir,
  pb_open,
  pl_repl_values
)
}
\arguments{
\item{ps_path}{path to the document to be created}

\item{ps_template}{name of the template}

\item{ps_package}{package containing the template}

\item{ps_create_dir}{flag whether document should be put into new directory}

\item{pb_open}{flag whether to open the newly created document in rstudio}

\item{pl_repl_values}{list of template placeholder replacement values}
}
\description{
This function calls \code{rmarkdown::draft()} to create a new document.
}
