% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_rmd.R
\name{get_default_create_dir}
\alias{get_default_create_dir}
\title{Find Default Option for create_dir}
\usage{
get_default_create_dir(ps_template, ps_package = NULL)
}
\arguments{
\item{ps_template}{name of the template}

\item{ps_package}{name of the package that contains the table}
}
\description{
Find the value of the option create_dir from the template.yaml in the
template directory
}
\details{
The template.yaml file is read using \code{rmarkdown:::yaml_load_file}
}
