% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_helper.R
\name{get_generic_replacement_values}
\alias{get_generic_replacement_values}
\title{Defaults For Generic Comparison Plot Report Template Replacement Values}
\usage{
get_generic_replacement_values()
}
\value{
list of comparison plot report replacement values
}
\description{
The template for the comparison plot report contains a number of placeholders
which are to be replaced by values when the report is generated. This function
returns useful default values to be inserted into the template as a replacement
for the placeholders.
}
