% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_helper.R
\name{get_repl_value_defaults_qprojektreport}
\alias{get_repl_value_defaults_qprojektreport}
\title{Defaults for Qualitas AG Projekt Report in Gernman}
\usage{
get_repl_value_defaults_qprojektreport()
}
\value{
list with replacement values
}
\description{
Replacement values that are specific for the format of a Qualitas AG
Projekt Report (in German) are returned.
}
