% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_helper.R
\name{get_repl_value_defaults_tufte}
\alias{get_repl_value_defaults_tufte}
\title{Defaults For Tufte Report Value Replacement List}
\usage{
get_repl_value_defaults_tufte()
}
\value{
default for tufte report value replacement list
}
\description{
The placeholder replacement in a RMarkdown template skeleton expectes a
fixed number of replacement values. The default values for this value
replacement list is returned
}
