% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_helper.R
\name{merge_list_to_default}
\alias{merge_list_to_default}
\title{Merge List To Default}
\usage{
merge_list_to_default(pl_base, pl_default)
}
\arguments{
\item{pl_base}{list of base entries}

\item{pl_default}{list with default entries}
}
\value{
l_result list with merged entries of pl_base and pl_default
}
\description{
Given two lists (pl_base and pl_default), the lists should be merged into a
result list (l_result) such that the result contains all entries of pl_base
and for the entries whose names are in pl_default, but are not in pl_base,
the entries of pl_default is included in l_result. This leads to a list (l_result)
which has the same names as pl_default, but for those names which are also
in pl_base, the values from pl_base overwrite the values of pl_default.
}
\details{
This merging procedure is used in replacement of placeholders in templates
where only parts of the placeholders are replaced with specified values and
the rest is replaced with assumed default values.
}
\examples{
\dontrun{
n <- 9
l2 <- lapply(1:n, function(x) x)
names(l2) <- LETTERS[1:9]
merge_list_to_default(pl_base = list(A = 10L, B = 20L), pl_default = l2)
}
}
