% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_helper.R
\name{string_replace}
\alias{string_replace}
\title{Replace Placeholders in Template with Replacement Values}
\usage{
string_replace(
  ps_tmpl,
  pl_repl_value,
  ps_ph_open_tag = "<ph>",
  ps_ph_end_tag = "</ph>"
)
}
\arguments{
\item{ps_tmpl}{template text containing placeholders}

\item{pl_repl_value}{list with replacement values}

\item{ps_ph_open_tag}{open-tag for placeholder}

\item{ps_ph_end_tag}{end-tag for placeholder}
}
\value{
s_result where placeholders have been replaced with replacement values
}
\description{
In a template text with placeholders, these placeholders are replaced with
values that are specified via the list of replacement values. As a result
the text is returned with the placeholders replaced by the values specified
in the list of replacement values.
}
\details{
Placeholders are words that are enclosed between an opening tag and an end tag.
These tags can be specified as arguments ps_ph_open_tag and ps_ph_end_tag.
The word between the tags corresponds to the name of the placeholder. This
name is also used as a name in the list of replacement values. The value
with which the placeholder is to be replaced is given as the list element
associated with the name of the placeholder.
}
