% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_helper.R
\name{sub_pattern_replacement}
\alias{sub_pattern_replacement}
\title{Replacement of Placeholders}
\usage{
sub_pattern_replacement(
  ps_path,
  pl_repl_value,
  ps_ph_open_tag = "<ph>",
  ps_ph_end_tag = "</ph>"
)
}
\arguments{
\item{ps_path}{Path with template text}

\item{pl_repl_value}{list of placeholder replacement values}

\item{ps_ph_open_tag}{opening tag for placeholder}

\item{ps_ph_end_tag}{closing tag for placeholder}
}
\description{
Replacement of Placeholders
}
